// apps/web/src/data/solar/elements.gen.ts
export const SATELLITES = {
 "epoch": 2461309.57811,
 "count": 14420,
 "counts": [
  22,
  11133,
  651,
  1854,
  169,
  572,
  19
 ],
 "tles": {
  "iss": {
   "norad": 25544,
   "line1": "1 25544U 98067A   26268.17397659  .00011139  00000+0  20797-3 0  9993",
   "line2": "2 25544  51.6316 165.2379 0004773 179.0895 181.0103 15.49282084587251"
  },
  "tiangong": {
   "norad": 48274,
   "line1": "1 48274U 21035A   26268.14703992  .00015428  00000+0  18724-3 0  9999",
   "line2": "2 48274  41.4679  75.0253 0002138 306.5316  53.5325 15.60268422308772"
  },
  "hubble": {
   "norad": 20580,
   "line1": "1 20580U 90037B   26266.94713825  .00005513  00000+0  16734-3 0  9992",
   "line2": "2 20580  28.4727 125.0661 0001663 151.6515 208.4172 15.31734421803723"
  }
 },
 "fetched": "2026-09-25"
} as const;